refactor: extract shared FanSyncOptimisticEntity base for fan/light
fan.py and light.py carried ~150 lines of near-identical optimistic-update
machinery (per-key overlay, guard window, confirm-via-push-or-poll, revert
on failure). Extract it into a new FanSyncOptimisticEntity base in entity.py;
fan/light now subclass it and supply only OVERLAY_KEYS, their HA-specific
state properties/commands, and a _log_state hook.

Also unifies the confirm-predicate convention: predicates always receive this
device's per-device status (previously fan passed per-device status while
light passed the full aggregated dict and unwrapped internally), and tightens
the moved type annotations (dict[str, object] / Callable[[dict[str, object]],
bool]) plus parenthesizes a multi-except.

Behavior-preserving: full suite unchanged (185 passed). The snapback tests
patch CONFIRM_INITIAL_DELAY_SEC at its new home (entity) since that constant's
binding moved with the code.
