+++ tests/test_snapback_behavior.py
@@ -77,7 +77,7 @@ async def test_fan_snapback_after_guard_expiry(hass: HomeAssistant):
 
     # Initiate optimistic change to 55%
     with (
-        patch("custom_components.fansync.fan.CONFIRM_INITIAL_DELAY_SEC", 0),
+        patch("custom_components.fansync.entity.CONFIRM_INITIAL_DELAY_SEC", 0),
         patch("custom_components.fansync.fan.time.monotonic", side_effect=lambda: fake_monotonic()),
     ):
         await hass.services.async_call(
@@ -110,7 +110,7 @@ async def test_light_snapback_after_guard_expiry(hass: HomeAssistant):
 
     # Initiate optimistic light on with brightness ~50%
     with (
-        patch("custom_components.fansync.light.CONFIRM_INITIAL_DELAY_SEC", 0),
+        patch("custom_components.fansync.entity.CONFIRM_INITIAL_DELAY_SEC", 0),
         patch(
             "custom_components.fansync.light.time.monotonic", side_effect=lambda: fake_monotonic()
         ),
